requirements.txt
# Note Agent Dependencies

# Config parsing
pyyaml>=6.0
